# swagger.yaml
openapi: 3.0.3
info:
  title: Hospital Backend API
  description: |
    Hospital Management System API for managing employee salary records, pharmacy data, and analytics.

    ## Features
    - JWT-based authentication with refresh token rotation
    - Excel/CSV file upload with Nepali language support
    - Employee salary record management
    - Pharmacy record management
    - Analytics and reporting with export capabilities

    ## Authentication
    Most endpoints require a valid JWT access token in the Authorization header.
    Access tokens expire after 15 minutes. Use the refresh token endpoint to obtain new tokens.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login or register
  schemas:
...
    PharmacyRecord:
      type: object
      properties:
        _id:
          type: string
          example: "507f1f77bcf86cd799439013"
        pan:
          type: string
          example: "123456789"
        name:
          type: string
          example: "Pharmacist Name"
        position:
          type: string
          example: "Senior Pharmacist"
        nightDutyCount:
          type: number
          example: 5
        rate:
          type: number
          example: 600
        grossAmount:
          type: number
          example: 15000
        taxDeduction:
          type: number
          example: 1500
        netPayable:
          type: number
          example: 13500
        accountNumber:
          type: string
          example: "1234567890123"
        source:
          type: string
          example: "pharmacy_data.xlsx"
        uploadedAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 50
        total:
          type: integer
          example: 100
        pages:
          type: integer
          example: 2
    SalaryTotals:
      type: object
      properties:
        totalGross:
          type: number
          example: 500000
        totalNet:
          type: number
          example: 450000
        totalTax:
          type: number
          example: 50000
    UploadFileResult:
      type: object
      properties:
        filename:
          type: string
          example: "salary_data.xlsx"
        rowsRead:
          type: integer
          example: 250
        inserted:
          type: integer
          example: 225
        skipped:
          type: integer
          example: 25
        errors:
          type: array
          items:
            type: string
          example: ["Row 5: Invalid PAN"]
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message description"
...
paths:
...
  /api/upload:
    post:
      tags:
        - Upload
      summary: Upload salary data files
      description: |
        Upload Excel or CSV files containing salary data. Supports:
        - Multiple files (max 10 files, 25MB each)
        - Excel (.xlsx, .xls) and CSV formats
        - Nepali language (Preeti font to Unicode conversion)
        - Flexible column header mapping
        - Deduplication via row hashing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Excel or CSV files to upload
      responses:
        '200':
          description: Files processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  filesProcessed:
                    type: integer
                    example: 2
                  totalRowsRead:
                    type: integer
                    example: 500
                  totalInserted:
                    type: integer
                    example: 450
                  totalSkipped:
                    type: integer
                    example: 50
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/UploadFileResult'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/pharmacy/upload:
    post:
      tags:
        - Pharmacy
      summary: Upload pharmacy salary data
      description: Upload Excel files containing pharmacy salary data
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Excel files to upload (max 10 files)
      responses:
        '200':
          description: Files processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  filesProcessed:
                    type: integer
                    example: 1
                  totalRowsRead:
                    type: integer
                    example: 100
                  totalInserted:
                    type: integer
                    example: 95
                  totalUpdated:
                    type: integer
                    example: 5
                  totalSkipped:
                    type: integer
                    example: 0
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        filename:
                          type: string
                        rowsRead:
                          type: integer
                        inserted:
                          type: integer
                        updated:
                          type: integer
                        skipped:
                          type: integer
                        errors:
                          type: array
                          items:
                            type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/pharmacy:
    get:
      tags:
        - Pharmacy
      summary: List pharmacy records
      description: Get paginated list of pharmacy records with totals
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Results per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of pharmacy records
          content:
            application/json:
              schema:
                type: object
                properties:
                  records:
                    type: array
                    items:
                      $ref: '#/components/schemas/PharmacyRecord'
                  totals:
                    $ref: '#/components/schemas/SalaryTotals'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...